Extract character and script into new files.
