+++ __init__.py
@@ -1,0 +1,3 @@
+from .bot import Game, Player, Townsfolk, Outsider, Minion, Demon, Traveler, AbilityModifier, NULL_GAME
+
+__all__ = ['Game', 'Player', 'Townsfolk', 'Outsider', 'Minion', 'Demon', 'Traveler', 'AbilityModifier', 'NULL_GAME']
