+++ model/__init__.py
@@ -1,0 +1,4 @@
+from .character import Character
+from .script import Script
+
+__all__ = ['Character', "Script"]
